:bento:TP3: using copilot  :art:

+++ bibliotheque-semaine18/app/Http/Controllers/AuthorController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\StoreAuthorRequest;
 use App\Models\Author;
 use Illuminate\Http\Request;
 
@@ -9,6 +10,8 @@ class AuthorController extends Controller
 {
     public function index()
     {
+        $this->authorize('viewAny', Author::class);
+
         $authors = Author::all();
 
         return view('authors.index', compact('authors'));
@@ -16,44 +19,40 @@ class AuthorController extends Controller
 
     public function create()
     {
+        $this->authorize('create', Author::class);
+
         return view('authors.create');
     }
 
-    public function store(Request $request)
+    public function store(StoreAuthorRequest $request)
     {
-        $request->validate([
-            'last_name' => 'required',
-            'first_name' => 'required',
-            'email' => 'required|email|unique:authors',
-            'phone' => 'nullable',
-        ]);
+        $this->authorize('create', Author::class);
 
-        Author::create($request->all());
+        Author::create($request->validated());
 
         return redirect()->route('authors.index');
     }
 
     public function edit(Author $author)
     {
+        $this->authorize('update', $author);
+
         return view('authors.edit', compact('author'));
     }
 
-    public function update(Request $request, Author $author)
+    public function update(StoreAuthorRequest $request, Author $author)
     {
-        $request->validate([
-            'last_name' => 'required',
-            'first_name' => 'required',
-            'email' => 'required|email|unique:authors,email,' . $author->id,
-            'phone' => 'nullable',
-        ]);
+        $this->authorize('update', $author);
 
-        $author->update($request->all());
+        $author->update($request->validated());
 
         return redirect()->route('authors.index');
     }
 
     public function destroy(Author $author)
     {
+        $this->authorize('delete', $author);
+
         $author->delete();
 
         return redirect()->route('authors.index');
